agents/slave-agent/question-answering/requirements.txt
# Base dependencies
python-dotenv
requests
google-generativeai
google-genai
pydantic
pydantic-settings
trento-agent-sdk

# For the web server
fastapi
uvicorn[standard]

# Utilities
scikit-learn
torch

# For healthchecks and monitoring
httpx

# Development tools
pytest
black
isort
mypy
pylint